Move MarkerTree into its own file
